Begin refactor
