+++ packages/annotation-ui/src/index.js
@@ -1,8 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
-import Store from "./model/Store";
+import AppShell from "./AppShell";
+import Store from "./model";
+import { Layers } from "./Task";
 
 import "react-mosaic-component/react-mosaic-component.css";
 import "@blueprintjs/core/lib/css/blueprint.css";
@@ -11,7 +12,7 @@ import "@blueprintjs/icons/lib/css/blueprint-icons.css";
 ReactDOM.render(
   <React.StrictMode>
     <Store>
-      <App />
+      <AppShell layers={() => <Layers />} />
     </Store>
   </React.StrictMode>,
   document.getElementById("app")
