+++ packages/annotation-ui/src/model/Store.js
@@ -4,11 +4,6 @@ import lodash_get from "lodash.get";
 import lodash_unset from "lodash.unset";
 
 const initialState = {
-  init: {
-    srcVideo: "/cdn/test.mp4",
-    vidHeight: 360,
-    vidWidth: 480,
-  },
   selectedLayer: ["Query 1"],
   debug: {
     actionsFired: [],
