+++ packages/annotation-ui/src/model/index.js
@@ -1,0 +1,4 @@
+// TODO: Make this its own component
+import Store from "./Store";
+export * from "./Store";
+export default Store;
