+++ packages/annotation-ui/src/panels/LayersPanel.js
@@ -1,14 +1,13 @@
 import React from "react";
-import { Layers } from "../Task";
 import DebugPanel from "./DebugPanel";
 
-export default function ContextPanel() {
+export default function LayersPanel(props) {
   return (
     <div style={{ display: "flex", height: "100%", flexDirection: "column" }}>
       <div style={{ flex: 1 }}>
         <div className="bp3-tree">
           <ul className="bp3-tree-node-list bp3-tree-root">
-            <Layers />
+            <props.layers />
           </ul>
         </div>
       </div>
